move hasher to cryptography
