--- hasher/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "commonware-hashing"
-edition = "2021"
-publish = true
-version = "0.0.2"
-license = "MIT OR Apache-2.0"
-description = "Leverage common functionality across multiple primitives."
-readme = "README.md"
-homepage = "https://commonware.xyz"
-repository = "https://github.com/commonwarexyz/monorepo/tree/main/utils"
-documentation = "https://docs.rs/commonware-utils"
-
-[dependencies]
-bytes = { workspace = true }
-sha2 = { workspace = true }
